apq8084: sensors: proximity: fix reading calibration data from file
* The number of bytes is checked against the wrong size value.

+++ drivers/sensorhub/atmel/factory/prox_tmg399x.c
@@ -214,7 +214,7 @@ int proximity_open_calibration(struct ssp_data *data)
 
 	iRet = cancel_filp->f_op->read(cancel_filp,
 		(u8 *)&data->uProxCanc, sizeof(unsigned int), &cancel_filp->f_pos);
-	if (iRet != sizeof(u8)) {
+	if (iRet != sizeof(unsigned int)) {
 		pr_err("[SSP]: %s - Can't read the cancel data\n", __func__);
 		iRet = -EIO;
 	}
